validate user

+++ controller/userController.js
@@ -44,19 +44,67 @@ async function getUserById(req, res) {
 
 async function editUser(req, res) {
     try {
-        const { username } = req.body;
+        const { username,password } = req.body;
         const id = req.params.id;
 
-        await users.update({
-            username
-        }, {
-            where: { id }
-        })
 
-        res.status(200).json({
-            status: 'success',
-            message: `data dari id ${id} nya berhasil berubah`
-        })
+        const checkRequired = req.body.username && req.body.password ? true : false
+        const checkPassword = /^(?=.*\d)(?=.*[a-z])(?=.*[A-Z]).{6,20}$/;
+        const checkId = await users.findOne({where: {id: id}}) 
+        const checkUsername = await users.findOne({where: {username: username}}) 
+
+        if (password && username) {
+            if (!checkRequired){
+                res.status(400).json({
+                    status: 'failed',
+                    message: "Data tidak lengkap"
+                })
+            }else if (!password.match(checkPassword)){
+                res.status(400).json({
+                    status: 'failed',
+                    message: "Password tidak sesuai harap terdapat 1 digit, 1 lowercase , 1 uppercase dan dengan panjang 6-10"
+                })
+            }else if (checkId.id !== checkUsername.id){
+                res.status(400).json({
+                    status: 'failed',
+                    message: "username sudah ada"
+                })
+            }else{
+                const hashPassword = bcrypt.hashSync(password, 10)
+                await users.update({
+                    username,
+                    password : hashPassword
+                }, {
+                    where: { id }
+                })
+        
+                res.status(200).json({
+                    status: 'success',
+                    message: `data dari id ${id} nya berhasil berubah`
+                })
+            }
+        }else{
+
+            if (checkId.id !== checkUsername.id){
+                res.status(400).json({
+                    status: 'failed',
+                    message: "username sudah ada"
+                })
+            }else{
+                await users.update({
+                    username
+                }, {
+                    where: { id }
+                })
+        
+                res.status(200).json({
+                    status: 'success',
+                    message: `data dari id ${id} nya berhasil berubah 2`
+                })
+            }
+        }
+
+        
     } catch (err) {
         res.status(400).json({
             status: 'failed',
@@ -67,19 +115,28 @@ async function editUser(req, res) {
 
 async function deleteUser(req, res) {
     try {
-        const id = req.params.id
-        await users.destroy({
-            where: {
-                id
-            }
-        })
+        const id = await users.findOne({where : {id : req.params.id}})
+        if (!id){
+            res.status(400).json({
+                "message" : "data tidak ada"
+            })
+        }else{
+            await users.destroy({
+                where: {
+                    id : req.params.id
+                }
+            })
+            res.status(200).json({
+                'status': 'success',
+                'message': `data ${req.params.id} ini berhasil di hapus`
+            })
+        }
 
-        res.status(200).json({
-            'status': 'success',
-            'message': `data ${id} ini berhasil di hapus`
-        })
     } catch (err) {
-        res.status(400).message(err.message)
+        res.status(400).json({
+            status: 'failed',
+            message: err.message
+        })
     }
 }
 
@@ -90,8 +147,7 @@ async function createUser(req, res) {
 
         const checkRequired = req.body.username && req.body.password ? true : false
         const checkPassword = /^(?=.*\d)(?=.*[a-z])(?=.*[A-Z]).{6,20}$/;
-        const checkUsername = users.findIndex((e) => e.name)
-        console.log(checkRequired)
+        const checkUsername = await users.findOne({where: {username: username}}) 
         if (!checkRequired){
             res.status(400).json({
                 status: 'failed',
@@ -102,19 +158,26 @@ async function createUser(req, res) {
                 status: 'failed',
                 message: "Password tidak sesuai harap terdapat 1 digit, 1 lowercase , 1 uppercase dan dengan panjang 6-10"
             })
+        }else if (checkUsername){
+            res.status(400).json({
+                status: 'failed',
+                message: "username sudah ada"
+            })
+        }else{
+            const hashPassword = bcrypt.hashSync(password, 10)
+            const newUser = await users.create({
+                username,
+                password : hashPassword,
+            })
+            res.status(201).json({
+                status: 'success',
+                data: {
+                    user: newUser
+                }
+            })
         }
 
-        const hashPassword = bcrypt.hashSync(password, 10)
-        const newUser = await users.create({
-            username,
-            password : hashPassword,
-        })
-        res.status(201).json({
-            status: 'success',
-            data: {
-                user: newUser
-            }
-        })
+        
     } catch (err) {
         res.status(400).json({
             status: 'failed',
